[FLINK-964] tear apart core plotting code and configuration
